requirements.txt
numpy 
pandas
python-dotenv  # for reading environment variables .env file 
mysql-connector-python
pymysql
scikit-learn
seaborn
ipykernel
catboost 
xgboost
mlflow